Refactor compare runner metrics helpers
